refining response models and errors handlers on team and user routers

+++ app/routers/team.py
@@ -1,36 +1,41 @@
 from fastapi.responses import JSONResponse
 from fastapi.encoders import jsonable_encoder
-from fastapi import APIRouter,Depends, HTTPException, status
-from typing import List, ForwardRef
-from ..database.database import SessionLocal
+from fastapi import APIRouter,Depends, HTTPException, Response, status
+# from typing import  ForwardRef
+from app.database.database import SessionLocal
 from sqlalchemy.orm import Session
-from ..models.user import User as UserModel
-from ..schemas.user_team import TeamUpdateSchema,TeamBase,Team as TeamSchema
-from ..models.team  import Team
-from ..models.user_team import User_team
-from ..schemas.user_team import Member
-from ..schemas.user_team import User
-from ..controllers.userController import require_role,get_current_active_user
-from ..schemas.invitacion import Invitacion,Invitacion_Response
-from ..models.invitacion import Invitacion as InvitacionModel
+from app.models.user import User as UserModel
+from app.schemas.user_team import TeamCreate, TeamUpdateSchema,TeamBase,User,Member,TeamReturn
+from app.models.team  import Team
+from app.models.user_team import User_team
+from app.controllers.userController import require_role,get_current_active_user
+from app.schemas.invitacion import InvitacionCreate,InvitacionResponse
+from app.models.invitacion import Invitacion as InvitacionModel
 import json
 from uuid import uuid4
-from ..controllers.teamController import add_user_to_team
+from app.controllers.teamController import add_user_to_team
 from datetime import datetime,timedelta
 from typing import Annotated
 
 
-teamRouter = APIRouter(prefix="/Team",tags=["Teams"])
+teamRouter = APIRouter(prefix="/Teams",tags=["Teams"])
 def get_db():
     db = SessionLocal()
     try:
         yield db
     finally:
         db.close()
-@teamRouter.post("/create_team")
-async def create_team(team: TeamBase ,user:User=Depends(require_role("Admin")),db: Session = Depends(get_db)):
+@teamRouter.post("/",response_model=TeamReturn)
+async def create_team(team: TeamCreate ,user:User=Depends(require_role("Admin")),db: Session = Depends(get_db)):
     try:
+        session_team = db.query(Team).filter(Team.name == team.name).first()
+
         # Player =UserModel(email="Daniel",password="1248")
+        if session_team:
+            raise HTTPException(
+            status_code=400,
+            detail="The team with this name already exists in the system",
+        ) 
         teamModel =Team(**team.model_dump())
         teamModel.creator_id = user.id
         print(teamModel)
@@ -41,11 +46,12 @@ async def create_team(team: TeamBase ,user:User=Depends(require_role("Admin")),d
         print(userteam)
         db.add(userteam)
         db.commit()
-        team_schema = TeamSchema(name=teamModel.name,creator=user,id=teamModel.id,members=[])
-        serializable = team_schema.model_dump()
-        print(serializable)
-        response = {"Team": serializable}
-        return JSONResponse(content=response,status_code=200)
+        return teamModel
+        # team_schema = TeamReturn(name=teamModel.name,creator=user,id=teamModel.id,members=[])
+        # serializable = team_schema.model_dump()
+        # print(serializable)
+        # response = {"Team": serializable}
+        # return JSONResponse(content=response,status_code=200)
     except Exception as e:
         raise e
 # @teamRouter.post("/add_user_team",description="Agregar un miembro a un equipo",response_class= JSONResponse,)
@@ -54,37 +60,41 @@ async def create_team(team: TeamBase ,user:User=Depends(require_role("Admin")),d
 #         return JSONResponse(content= {"members":members["ids"]})
    
     
-@teamRouter.delete("/delete_team/{team_id}",description="Delete a team")
+@teamRouter.delete("/{team_id}",description="Delete a team")
 async def delete_team(team_id:int,db:Session = Depends(get_db),user:User =Depends(get_current_active_user)):
 
     team = db.query(Team).filter(Team.id == team_id).first()
+    if not team:
+        raise HTTPException(detail="This team doesn't exist",status_code=404)
     if team.creator_id == user.id:
         db.delete(team)
         db.commit()
     else:  
         raise HTTPException(status_code=401)
           
-    return JSONResponse(content="Deleted!",status_code=204)
+    return JSONResponse(content="Deleted!",status_code=200)
 
 
 
-@teamRouter.post("/Invite_user",description="Invite a user to the team")
-async def invite_user(invitacion:Invitacion,db: Session =Depends(get_db)):   
+@teamRouter.post("/Invite_user",description="Invite a user to the team",response_model=InvitacionResponse)
+async def invite_user(invitacion:InvitacionCreate,db: Session =Depends(get_db)):   
     token = str(uuid4())
-    link = f"https://127.0.0.1:8000/Team/Acept_invitacion/{token}"
+    # link = f"https://127.0.0.1:8000/Team/Acept_invitacion/{token}"
     invitacionModel =InvitacionModel(token=token,invited_user_id=invitacion.invited_user_id,team_id=invitacion.team_id)
     db.add(invitacionModel)
     db.commit()
-    Schemainvitacion = Invitacion_Response(token=invitacionModel.token,team_id=invitacionModel.team_id,invited_user_id=invitacionModel.invited_user_id)
-    dictInvitacion= Schemainvitacion.model_dump()
-    response_invitacion ={"Invitacion": dictInvitacion,"link":link}
+    db.refresh(invitacionModel)
+    # Schemainvitacion = Invitacion_Response(token=invitacionModel.token,team_id=invitacionModel.team_id,invited_user_id=invitacionModel.invited_user_id)
+  
+    # dictInvitacion= Schemainvitacion.model_dump()
+    # response_invitacion ={"Invitacion": dictInvitacion,"link":link}
     
-    return JSONResponse(content=response_invitacion,status_code=200)
+    return invitacionModel
 
 
 
     # print(teamModel.creator.email)
-@teamRouter.post("/Acept_invitacion/{token}")
+@teamRouter.post("/Acept_invitacion/{token}",response_model=list[User])
 async def Acept_invitacion(token:str,db:Session =Depends(get_db)):
     invitacion = db.query(InvitacionModel).filter_by(token=token).first()
     now = datetime.now()
@@ -97,17 +107,19 @@ async def Acept_invitacion(token:str,db:Session =Depends(get_db)):
     # Si el token es válido y la invitación está pendiente
     if invitacion and invitacion.status == "pendiente" and bool_token:
         # Añade al usuario al equipo
-        members = await add_user_to_team(new_member=Member(user_id=invitacion.invited_user_id,team_id=invitacion.team_id),db=db)
-        dict=[]
-        for member in members:
-            member_serialized = User(**member.__dict__)
-            member_serialized = member_serialized.model_dump()
-            dict.append(member_serialized)
+        team = await add_user_to_team(new_member=Member(user_id=invitacion.invited_user_id,team_id=invitacion.team_id),db=db)
+        # esta complejidad no es necesaria,solo fue para poder devovler una JsonResponse manualmente
+        # dict=[]
+        # for member in members:
+        #     member_serialized = User(**member.__dict__)
+        #     member_serialized = member_serialized.model_dump()
+        #     dict.append(member_serialized)
         # members_serialized = [User(**member.__dict__)for member in members]
-        print(member_serialized)
+        # print(member_serialized)
         invitacion.status = "Aceptada"
         db.commit()
-        return JSONResponse(content={"Members":dict},status_code=201)
+        # return JSONResponse(content={"Members":dict},status_code=201)
+        return team.members
     else:
         raise HTTPException(status_code=404, detail="Invitación no válida o expirada")
     
@@ -115,35 +127,45 @@ async def Acept_invitacion(token:str,db:Session =Depends(get_db)):
 async def drop_member(tarjet_member_id:int,id_team:int,user:Annotated[User, Depends(require_role("Admin"))],db:Session = Depends(get_db)):
     try:
         user_model:UserModel = db.query(UserModel).filter(UserModel.id == user.id).first()
+        team_model:Team =db.query(Team).filter(Team.id == id_team).first()
+        userteam= db.query(User_team).filter(User_team.user_id==tarjet_member_id).first()
+        if not team_model:
+            raise HTTPException(details="Team not found",status_code=404)
+        if not userteam:
+                raise HTTPException(details={"Info": f"User {tarjet_member_id} not found in team {id_team}"},status_code=404)
         print(user_model.teams_created)
-        for team in user_model.teams_created:
-            if id_team == team.id:
-                userteam= db.query(User_team).filter(User_team.user_id==tarjet_member_id).first()
-                print(userteam)
-                if not userteam:
-                    return JSONResponse(content={"Info": f"No se ha encontrado el usuario  {tarjet_member_id} en el equipo  {id_team}"})
-                db.delete(userteam)
-                db.commit()
-                return JSONResponse(content={"Info": f"Se ha expulsado el usuario{tarjet_member_id}"})
+        print(team_model)
+        if not team_model in user_model.teams_created:
+            raise HTTPException(status_code=430,detail="Operation not permitted,Not enough permissions(only team's owners are able to delete a member)") 
+        db.delete(userteam)
+        db.commit()
+        return JSONResponse(content={"Info": f"Se ha expulsado el usuario{tarjet_member_id}"})
+        # for team in user_model.teams_created:
+        #     if id_team == team.id:
+        #         userteam= db.query(User_team).filter(User_team.user_id==tarjet_member_id).first()
+        #         print(userteam)
+        #         if not userteam:
+        #             return JSONResponse(content={"Info": f"No se ha encontrado el usuario  {tarjet_member_id} en el equipo  {id_team}"})
+        #         db.delete(userteam)
+        #         db.commit()
+        #         return JSONResponse(content={"Info": f"Se ha expulsado el usuario{tarjet_member_id}"})
         # raise HTTPException(status_code=430,detail="Operation not permitted,Not enough permissions(should be a team manager)")   
     except Exception as e:
         raise e
-@teamRouter.get("/Get_Teams")
+@teamRouter.get("/",response_model=list[TeamReturn])
 async def get_teams(user:User=Depends(get_current_active_user),db:Session=Depends(get_db)):
     teams = db.query(Team).all()
     print(teams)
-    team_schema = [
-        {
-            "Team": TeamSchema(**team.__dict__).model_dump(),
-            "Tournaments_in": jsonable_encoder(team.tournaments)
-        }
-        for team in teams
-    ]
-    # print(type(team_schema[0]["Team"]))
-    # print(team_schema[0]["Tournaments_in"])
-    # return team_schema
-    return JSONResponse(content={"Teams":team_schema},status_code=200)
-    # return team_schema
+    # team_schema = [
+    #     {
+    #         "Team": TeamReturn(**team.__dict__).model_dump(),
+    #         "Tournaments_in": jsonable_encoder(team.tournaments)
+    #     }
+    #     for team in teams
+    # ]
+    # return JSONResponse(content={"Teams":team_schema},status_code=200)
+    return teams
+
     
 @teamRouter.patch("/update_team")
 async def update_team(teamSchema:TeamUpdateSchema,user:User=Depends(require_role("Admin")),db:Session =Depends(get_db)):
@@ -160,8 +182,8 @@ async def update_team(teamSchema:TeamUpdateSchema,user:User=Depends(require_role
     db.commit()
     db.refresh(team_to_update)
     # dicted= team_to_update.__dict__
-    response_team = TeamSchema.model_validate(team_to_update)
-    # response_team = TeamSchema(**team_to_update.__dict__)
+    response_team = TeamReturn.model_validate(team_to_update)
+    # response_team = TeamReturn(**team_to_update.__dict__)
     # print(dicted)
     print(response_team)
     return JSONResponse(content=response_team.model_dump(),status_code=200)
